Improvements to code assist in the REPL
Re-enable acronmyn-style completion, e.g. getClass.gdm`
offers `getDeclaredMethod[s]`.

Disable the typo-matcher in JLine which tends to offer confusing

Fix completion of keywored-starting-idents (e.g. `this.for<TAB>` offers
`formatted`.

Register a widget on CTRL-SHIFT-T that prints the type of the
expression at the cursor. A second invokation prints the
desugared AST.