JAR reading: option to use zipfs, other fixes
Multithreaded users of the compiler can experience lock contention
in classpath reading. This occurs because a) scalac by default reuses
a ZipFile instance for a given JAR across concurrent compilers to
reduce the memory footprint, and b) because j.u.ZipFile contains
does not support concurrent data and metadata reads.

The JDK does ship with ZipFileSystemProvider, however, that does
support concurrent readers. This PR adds an compiler option,
-Yzipfs-classpath, to use this in the classpath implementation.

I extended FileBasedCache to support extra data in the cache key,
such as the value of this setting. I noticed that the -release
setting should also form part of the cache key and added it.

Profiling showed my that the attempt to reduce allocations when
converting the classpath entry names into symbol names was counter
productive as it bypassed `lazy val name` in `ClassFileEntryImpl`
which was forced anyway on other code paths.